Check for Jupyter Notebook before allocating other terminal sizes.
Currently, the number of columns and rows used by `progressbar2` in Jupyter notebooks is inherited from the shell from which it was run... which leads to some pretty weird behaviour. Instead, I think we should check for Jupyter notebooks first.

Is there ever a case where this doesn't make sense?