Make printing a newline on finish optional
Sometimes, like if the progressbar is used as part of logging, it is useful to be able to overwrite the progress bar simply using `\r`, which is made more difficult by this dangling newline. This PR makes that optional (the keyword argument name can be changed... chosen here to match the `print` function).