Do not send garbage when the output is not a terminals
---

The problem:
------------

The progressbars are beautiful on an interactive shell.
However, when running from a CI or something else that produces a log file, it is just a big messy blob of garbage.

Depending on how `\r` is handled by the viewer, we get either a very long and noisy line of text, or 3 empty lines between each update.

We also get too many updates for a log file -- This is addressed on #206.

---

The solution:
-------------

It now detects if it is running on an interactive shell via `fd.isatty()`.

This detection is not completely reliable, so it can also be specified via argument or environment variables

If it is not an interactive terminal:
- It writes one update per line, instead of overwrite the existing line
- It removes ANSI color escapes.