Avoid flickering
StdRedirectMixin makes the bar erase itself before writing the new update.
This works fine and is invisible most of the time.

Most of the time -- If the refreshes are very fast, flickering becomes visible

This PR first checks if there is something buffered to be written in stdout/stderr, and doesn't erase the current line otherwise